requirements.txt
aiogram>=3.13.1
pyrogram>=2.0.106
tgcrypto>=1.2.5
aiosqlite>=0.20.0
python-dotenv>=1.0.1
qrcode>=7.4.2
pillow>=10.0.0
